requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
charset-normalizer==3.1.0
discord==2.2.2
discord.py==2.2.2
distlib==0.3.6
filelock==3.11.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.4
packaging==23.0
pkgconfig==1.5.5
platformdirs==3.2.0
psutil==5.9.4
virtualenv==20.21.0
virtualfish==2.5.5
yarl==1.8.2
openai~=0.27.4
pinecone-client~=2.2.1
discord-py-interactions~=5.1.0
web3~=6.2.0
ens-normalize~=2.0.0
pycryptodome~=3.17
python-dotenv~=1.0.0